pyproject.toml
[tool.poetry]
name = "test_task_3dlook"
version = "0.1.0"
description = "This is a Django project that provides API endpoints for managing products. The project implements a Product model with fields for name, description, uuid, created, updated, logo, and rotate_duration."
authors = ["Hikmat Samadov <s3m3dov@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
uuid = "^1.30"
boto3 = "^1.26.84"
django-storages = "^1.13.2"
Pillow = "^9.4.0"
celery = "5.1"
redis = "^4.5.1"
django-environ = "^0.10.0"
colorlog = "^6.7.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
